+++ app/experimenter/nimbus-ui/src/components/AppLayoutWithSidebar/index.stories.tsx
@@ -6,7 +6,12 @@ import { withLinks } from "@storybook/addon-links";
 import React from "react";
 import AppLayoutWithSidebar from ".";
 import { SERVER_ERRORS } from "../../lib/constants";
-import { NimbusExperimentStatus } from "../../types/globalTypes";
+import { mockChangelog } from "../../lib/mocks";
+import { getExperiment_experimentBySlug } from "../../types/getExperiment";
+import {
+  NimbusExperimentPublishStatus,
+  NimbusExperimentStatus,
+} from "../../types/globalTypes";
 import { Subject } from "./mocks";
 
 export default {
@@ -15,32 +20,63 @@ export default {
   decorators: [withLinks],
 };
 
-export const DraftStatus = () => (
-  <Subject
-    experiment={{
-      status: NimbusExperimentStatus.DRAFT,
-    }}
-  />
-);
+const storyWithExperimentProps = (
+  props: Partial<getExperiment_experimentBySlug>,
+  storyName?: string,
+) => {
+  const story = () => <Subject experiment={props} />;
+  story.storyName = storyName;
+  return story;
+};
 
-export const PreviewStatus = () => (
-  <Subject
-    experiment={{
-      status: NimbusExperimentStatus.PREVIEW,
-    }}
-  />
-);
-
-export const MissingDetails = () => (
-  <Subject
-    experiment={{
-      readyForReview: {
-        ready: false,
-        message: {
-          reference_branch: [SERVER_ERRORS.NULL_FIELD],
-          channel: [SERVER_ERRORS.EMPTY_LIST],
-        },
+export const MissingDetails = storyWithExperimentProps(
+  {
+    readyForReview: {
+      ready: false,
+      message: {
+        reference_branch: [SERVER_ERRORS.NULL_FIELD],
+        channel: [SERVER_ERRORS.EMPTY_LIST],
       },
-    }}
-  />
+    },
+  },
+  "Draft status, missing details",
+);
+
+export const DraftStatus = storyWithExperimentProps(
+  {
+    status: NimbusExperimentStatus.DRAFT,
+  },
+  "Draft status, filled out",
+);
+
+export const PreviewStatus = storyWithExperimentProps({
+  status: NimbusExperimentStatus.PREVIEW,
+});
+
+export const ReviewRequestedCannotReview = storyWithExperimentProps(
+  {
+    status: NimbusExperimentStatus.DRAFT,
+    statusNext: NimbusExperimentStatus.LIVE,
+    publishStatus: NimbusExperimentPublishStatus.REVIEW,
+    reviewRequest: mockChangelog(),
+  },
+  "Request requiring approval was made, user cannot review",
+);
+
+export const ReviewRequestedCanReview = storyWithExperimentProps(
+  {
+    status: NimbusExperimentStatus.DRAFT,
+    statusNext: NimbusExperimentStatus.LIVE,
+    publishStatus: NimbusExperimentPublishStatus.REVIEW,
+    reviewRequest: mockChangelog(),
+    canReview: true,
+  },
+  "Request requiring approval was made, user can review",
+);
+
+export const LiveStatus = storyWithExperimentProps(
+  {
+    status: NimbusExperimentStatus.LIVE,
+  },
+  "Live status",
 );
