+++ app/experimenter/nimbus-ui/src/components/PageEditAudience/index.tsx
@@ -73,7 +73,7 @@ const PageEditAudience: React.FunctionComponent<RouteComponentProps> = () => {
           refetchReview.current!();
 
           if (next) {
-            navigate("../request-review");
+            navigate("../");
           }
         }
       } catch (error) {
