+++ app/experimenter/nimbus-ui/src/components/PageSummary/FormLaunchConfirmationCheckboxes.tsx
